requirements.txt
streamlit
openai
python-docx
pytesseract
Pillow
pymupdf
